+++ auth_sys/urls.py
@@ -8,9 +8,4 @@ urlpatterns = [
     path('signup/', SignUpView.as_view(), name='signup'),
     path('login/', CustomLoginView.as_view(), name='login'),
     path('logout/', logout_view, name='logout'),
-    path('portfolio_main/<int:pk>/', views.PortfolioMainDetailView.as_view(), name="portfolio_main"),
-    path('portfolio_create/', views.PortfolioCreateView.as_view(), name='portfolio_create'),
-    path('portfolio_detail/<int:pk>/',views.PortfolioDetailView.as_view(), name="portfolio_detail"),
-    path('projects_list/<int:pk>/', views.ProjectsInformationView.as_view(), name="projects_list"),
-    path('projects_create/', views.ProjectsCreateView.as_view(), name="projects_create"),
 ]
