+++ portfolio/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from portfolio.models import Portfolio, PortfolioProjects
+
+# Register your models here.
+admin.site.register(Portfolio)
+admin.site.register(PortfolioProjects)
